Take the grouped columnwise scale prefix in 64 bits
_store_grouped_col_sf_u32 addresses a group's swizzled columnwise scale tile by the
span of the preceding groups:

    prefix_words = hidden * group_start // cutlass.Int32(64)

Both operands are Int32, so the product wraps once it passes 2^31. At DeepSeek-V3 671B
(hidden 7168) that happens as soon as group_start reaches 299,593 rows: the multiply
goes negative, the quotient with it, and the scale word is stored far below the buffer.

The quotient itself is harmless -- at most hidden * tokens / 64, about 59M words at the
largest shape here -- so only the multiply needs widening, and the index arithmetic that
follows stays 32-bit.

The failure is quiet before it is loud. Corruption starts at the overflow, but the
wrapped address usually still lands inside another live allocation, so the kernel keeps
running with wrong scales; it only faults once the address happens to fall outside a
mapped segment. Measured on GB200 at E=8, hidden 7168, comparing the columnwise scale
output against the Triton backend per group -- the ~1.5% baseline is the pre-existing
difference between the two backends, present at every size:

    tokens    max(hidden * group_start)        sfd differing, before -> after
     32,768         205,520,896                    1.6%  ->  1.6%
    262,144       1,644,167,168                    1.5%  ->  1.5%
    360,448       2,260,729,856  over 2^31        12.7%  ->  1.8%
    458,752       2,877,292,544  over 2^31        crash  ->  1.7%

compute-sanitizer named the store directly, an invalid 4-byte __global__ write from
Tcgen05GroupRowColFused whose address sits 112,918,524 bytes *before* the nearest
allocation -- the negative offset being the tell. It now reports zero invalid accesses
at the shape that used to fault.

The 3-GEMM MoE forward at 524,288 rows, which is where this first showed up, runs at
50.3 ms against the Triton backend's 68.5 ms; it previously could not run at all.

No measurable cost: the multiply is once per work item, and grouped rht_quantize_row_col
over 512..8192 tokens/expert is unchanged at 1.29x, 1.74x, 1.81x, 1.87x, 1.92x over
Triton (1.28x, 1.73x, 1.81x, 1.86x, 1.91x before).

test_group_rht_quantize_row_col.py and test_nvfp4_grouped_mm.py: 85 passed. Neither
reaches the overflow -- their shapes are far below 299,593 rows -- so this is not
covered by a regression test; a test that allocates ~6 GB of activations would be needed
to reach it.

+++ torchao/prototype/moe_training/nvfp4_training/_cutedsl_group_kernels_impl.py
@@ -1628,7 +1628,14 @@ def _store_grouped_col_sf_u32(mSF_u32, rSF, r, c_base, g, offsets_t, hidden):
     r_grp = (r % cutlass.Int32(128)) // cutlass.Int32(32)
     # A group has ``hidden * group_len / 64`` u32 scale words.  The preceding
     # concatenated groups cover the same expression with ``group_start``.
-    prefix_words = hidden * group_start // cutlass.Int32(64)
+    # The product is taken in 64 bits: at DeepSeek-V3 671B (hidden 7168) it passes
+    # 2^31 once ``group_start`` reaches 299,593 rows, and an Int32 multiply wraps
+    # negative there, so the store lands far below the buffer. The quotient is at
+    # most ``hidden * tokens / 64``, which is comfortably Int32, so only the
+    # multiply needs widening and the index arithmetic below stays 32-bit.
+    prefix_words = cutlass.Int32(
+        cutlass.Int64(hidden) * cutlass.Int64(group_start) // cutlass.Int64(64)
+    )
     words_per_hidden_block = group_len * cutlass.Int32(2)
     c_local = c_base - group_start // cutlass.Int32(16)
     # Plain Python loops: this helper is not AST-preprocessed, so the trace
